refactor Color.java so that the constructor throws a IllegalArgumentException instead of the user defined exeption it did previosly.
reason:
when I was designing the system I was mistaken on the definitions of arguments and parameters and so I did not thing that this Build in error was appropriate. however after some research I have concluded that IllegalArgumentException is the right Exception to throw here as it will reduce code to maintain.

source:https://www.geeksforgeeks.org/argument-vs-parameter-in-java/#:~:text=When%20a%20function,called%20Formal%20Parameters